Updated the examples, added new CLI example
* Created new examples dir
* Moved old example to ./examples/flaskapp
* Added new example to ./cliapp
* updated van.py script for flaskapp to reflect new pyvan interface
