+++ examples/cliapp/van.py
@@ -1,0 +1,11 @@
+import pyvan
+
+
+OPTIONS = {
+    "main_file_name": "main.py",
+    "show_console": True,
+    "use_existing_requirements": True,
+}
+
+
+pyvan.build(**OPTIONS)
