+++ examples/flaskapp/static/scripts.js
