+++ examples/flaskapp/static/styles.css
