+++ examples/flaskapp/templates/index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <link rel="stylesheet" href="../static/metro-all.min.css">
+    <link rel="stylesheet" href="../static/styles.css">
+    <title>FlaskApp</title>
+</head>
+<body>
+    
+<h1>Flask Desktop Application</h1>
+
+
+
+<a href="\home" class="button">Go home</a>
+
+
+
+<script src="../static/jquery-3.4.1.min.js"></script>
+<script src="../static/metro.min.js"></script>
+<script src="../static/scripts.js"></script>
+
+
+</body>
+</html>
