+++ examples/flaskapp/van.py
@@ -1,0 +1,11 @@
+import pyvan
+
+
+OPTIONS = {
+    "main_file_name": "main.py",
+    "show_console": False,
+    "use_pipreqs": True
+}
+
+
+pyvan.build(**OPTIONS)
